[REF] Move some things in workflow_activity_action
