+++ workflow_activity_action/models/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
+from . import wkf_action_model
 from . import wkf_activity
-from . import task
-from . import wkf_workitem
