+++ workflow_activity_action/views/wkf_activity.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+	    <record id="workflow_activity_form" model="ir.ui.view">
+            <field name="name">workflow.activity.form (workflow_activity_action)</field>
+            <field name="model">workflow.activity</field>
+            <field name="inherit_id" ref="base.view_workflow_activity_form"/>
+            <field name="arch" type="xml">
+                <xpath expr="//group[group[@string='Actions']]" position="after">
+                    <group string="Action available in this activity">
+                        <field name="use_action_object" />
+                        <field name="action_domain" />
+                        <field name="condition" />
+                        <field name="action_group" />
+                        <field name="action_ids">
+                            <tree editable="bottom">
+                                <field name="name" />
+                                <field name="action" />
+                            </tree>
+                        </field>
+                    </group>
+                </xpath>
+             </field>
+        </record>
+    </data>
+</openerp>
