+++ workflow_activity_action/views/workflow_activity_action.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <template id="assets_backend" name="workflow_activity_action" inherit_id="web.assets_backend">
+            <xpath expr="." position="inside">
+                <script type="text/javascript" src="/workflow_activity_action/static/src/js/view_form.js"></script>
+            </xpath>
+        </template>
+    </data>
+</openerp>
