+++ workflow_task/views/wkf_activity.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
-	    <record id="workflow_activity" model="ir.ui.view">
-            <field name="name">workflow.activity.form</field>
+	    <record id="workflow_activity_form" model="ir.ui.view">
+            <field name="name">workflow.activity.form (workflow_task)</field>
             <field name="model">workflow.activity</field>
-            <field name="inherit_id" ref="base.view_workflow_activity_form"/>
+            <field name="inherit_id" ref="workflow_activity_action.workflow_activity_form"/>
             <field name="arch" type="xml">
                 <xpath expr="//group[group[@string='Actions']]" position="after">
                     <group string="Tasks">
@@ -19,19 +19,9 @@
                             <field name="critical_delay" attrs="{'invisible': [('task_deadline_days', '=', False)]}"/>
                         </group>
                     </group>
-                    <group string="Action available in this activity">
-                        <field name="use_action_object" />
-                        <field name="use_action_task" attrs="{'invisible': [('task_create', '=', False)]}"/>
-                        <field name="action_domain" />
-                        <field name="condition" />
-                        <field name="action_group" />
-                        <field name="action_ids">
-                            <tree editable="bottom">
-                                <field name="name" />
-                                <field name="action" />
-                            </tree>
-                        </field>
-                    </group>
+                </xpath>
+                <xpath expr="//field[@name='use_action_object']" position="after">
+                    <field name="use_action_task" attrs="{'invisible': [('task_create', '=', False)]}"/>
                 </xpath>
              </field>
         </record>
